backend/requirements.txt
Django
djangorestframework
python-dotenv
django-cors-headers
gunicorn
nltk
langchain
langchain-pinecone
langchain-community
langchain-groq
matplotlib
numpy
pinecone
pandas
pypdf
scikit-learn
transformers
Bio

# (torch w/ cuda) pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
